% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-attgts.R
\name{calculate_att_g_t}
\alias{calculate_att_g_t}
\title{Estimate ATT(g,t) point estimates}
\usage{
calculate_att_g_t(
  g_val,
  t_val,
  lookup_table,
  N_table,
  hetero_var,
  verbose = FALSE
)
}
\arguments{
\item{g_val}{Which group to estimate ATT for.}

\item{t_val}{Which time period to estimate ATT for.}

\item{lookup_table}{Dataset with "counting" process which reflects when a unit is
first switched  on.}

\item{N_table}{Dataset with number of individuals per time and group..}

\item{hetero_var}{Factor variable for heterogeneous treatment effects.}

\item{verbose}{Whether to return all subcomponents used in ATT calculation (for debugging primarily)}
}
\description{
Given a 'lookup table' of when someone first was treated, calculate
the ATTs of Callaway and Sant'Anna
}
