% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{calculate_influence_function}
\alias{calculate_influence_function}
\title{Calculate 'rc' style influence function}
\usage{
calculate_influence_function(
  g_val,
  t_val,
  hetero_val = NULL,
  hetero_var = NULL,
  lookup_indiv_table,
  verbose = FALSE,
  check = FALSE,
  prop_score_known = FALSE
)
}
\arguments{
\item{g_val}{Group in ATT(g, t)}

\item{t_val}{t in ATT(g, t)}

\item{hetero_val}{Level of factor variable to estimate IF for.}

\item{hetero_var}{Factor variable for heterogeneous treatment effects.}

\item{lookup_indiv_table}{"counting" binary data}

\item{verbose}{Whether to return intermediate outputs}

\item{check}{Internal debugging function}

\item{prop_score_known}{Is the propensity score known}
}
\description{
Calculate 'rc' style influence function
}
