% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{create_cluster_rademacher}
\alias{create_cluster_rademacher}
\title{Create Rademacher Matrix with Clusters}
\usage{
create_cluster_rademacher(N_unique_clusters, cluster_id, biters)
}
\arguments{
\item{N_unique_clusters}{Number of clusters in cluster vector}

\item{cluster_id}{Nx1 vector where N is number of indiv. Holds CONTIGUOUS cluster id.}

\item{biters}{Number of bootstrap iterations to draw}
}
\description{
Draw rademacher weights for K clusters and B iters, then convert
into B x N matrix where N is number of individuals.
}
