% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-data.R
\name{create_group_first_treat_dt}
\alias{create_group_first_treat_dt}
\title{Create 'Group' counting data

Group individual counting data by treatment group and aggregate by summing.}
\usage{
create_group_first_treat_dt(
  dt,
  y_var,
  group_var,
  t_levels,
  group_levels,
  weight_df = NULL
)
}
\arguments{
\item{dt}{Indiv "counting" data.}

\item{y_var}{What is the y variable called.}

\item{group_var}{What is the group variable called.}

\item{t_levels}{Unique levels of t to calculate over}

\item{group_levels}{Unique levels of g to calculate over}

\item{weight_df}{Weighting df}
}
\description{
Create 'Group' counting data

Group individual counting data by treatment group and aggregate by summing.
}
