% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-did.R
\name{estimate_did}
\alias{estimate_did}
\title{Estimate CS DiD with Binary, Absorbing Treatment}
\usage{
estimate_did(
  data,
  y_var,
  group_var,
  t_var,
  id_var,
  birth_var = NULL,
  cluster_var = NULL,
  weight_df = NULL,
  prop_score_known = FALSE,
  biter = 1000,
  n_cores = 8
)
}
\arguments{
\item{data}{Dataset to use for estimation}

\item{y_var}{Integer Y variable that denotes the \emph{first} time period an individual is vaccinated}

\item{group_var}{Treatment group, in the style of CS package, must be first time period group is treated}

\item{t_var}{Time period variable.}

\item{id_var}{Unique ID for individuals.}

\item{birth_var}{If a panel with children being born, what period are they born in.}

\item{cluster_var}{Variable to cluster SEs on.}

\item{weight_df}{Dataframe of weights.}

\item{prop_score_known}{Is the propensity score known or estimated}

\item{biter}{How many bootstrap iterations to use to calculate SEs}

\item{n_cores}{How many cores to use when bootstrapping}
}
\description{
Estimates CS DiD when Y data is binary and 'absorbing' (once you're vaccinated
you cannot become unvaccinated).
}
