% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-did.R
\name{estimate_event_study}
\alias{estimate_event_study}
\title{Estimate Event Study}
\usage{
estimate_event_study(
  att_df,
  inf_matrix,
  balance_e = NULL,
  y_var = att_g_t,
  biter = 1000,
  n_cores = 8,
  prop_score_known = FALSE,
  group_vector = NULL,
  cluster_id = NULL
)
}
\arguments{
\item{att_df}{Data.table of ATTs}

\item{inf_matrix}{Influence function matrix, N individuals by K ATTs}

\item{balance_e}{Balance event study for e time periods}

\item{y_var}{Column name in att_df}

\item{biter}{Bootstrap draws for inference}

\item{n_cores}{Number of cores for parallal processing}

\item{prop_score_known}{Propensity score known or to be estimated}

\item{group_vector}{Nx1 vector of group IDs}

\item{cluster_id}{Nx1 vector of cluster IDs}
}
\description{
Given a data.table of ATTs and an influence matrix, create event study estimates.
}
