% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-did.R
\name{estimate_group_average}
\alias{estimate_group_average}
\title{Estimate Group Average}
\usage{
estimate_group_average(
  att_df,
  inf_matrix,
  balance_e = NULL,
  y_var = att_g_t,
  biter = 1000,
  n_cores = 8,
  cluster_id = NULL,
  aggregate_fits = FALSE
)
}
\arguments{
\item{att_df}{Data.table of ATTs}

\item{inf_matrix}{Influence function matrix, N individuals by K ATTs}

\item{balance_e}{Balance event study for e time periods}

\item{y_var}{Column name in att_df}

\item{biter}{Bootstrap draws for inference}

\item{n_cores}{Number of cores for parallal processing}

\item{cluster_id}{Nx1 vector of cluster IDs}

\item{aggregate_fits}{Are multiple fits across covariate factors present in att_df?}
}
\description{
Given a data.table of ATTs and an influence matrix, create group level estimates.
}
\details{
Assumes propensity score is known.
}
