% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabdid-package.R
\docType{package}
\name{fabdid-package}
\alias{fabdid}
\alias{fabdid-package}
\title{fabdid: Fast Absorbing Binary DiD}
\description{
Calculates ATTs and performs inference using Callaway and Sant'Anna (2021) for binary, "absorbing" outcomes quickly and (relatively) efficiently. Absorbing outcomes are those where an individual's outcome switches on in period t and is "on" for all subsequent periods.
}
\author{
\strong{Maintainer}: Ed Jee \email{edjee@uchicago.edu}

}
\keyword{internal}
