% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{run_nested_multiplier_bootstrap}
\alias{run_nested_multiplier_bootstrap}
\title{Run District/Town Multiplier Bootstrap}
\usage{
run_nested_multiplier_bootstrap(
  inf.func,
  biters,
  pl = FALSE,
  cores = 1,
  cluster_id_2 = NULL,
  weight_type = "rademacher"
)
}
\arguments{
\item{inf.func}{Influence function matrix. NxK where N is N indiv and K is # of ATTs}

\item{biters}{Bootstrap iterations. N.B. plural use of arg here but biter elsewhere. Nice.}

\item{pl}{Process in parallel}

\item{cores}{Number of cores to use if parallel processing}

\item{cluster_id_2}{"Upper" cluster id for nested clusters}

\item{weight_type}{\code{rademacher} or \code{rnorm} - which weight type to use

TAKEN DIRECTLY FROM BCALLAWAY11/DID - edited for nested booting}
}
\description{
Multiplier Bootstrap
}
