% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{run_multiplier_bootstrap}
\alias{run_multiplier_bootstrap}
\title{TAKEN DIRECTLY FROM BCALLAWAY11/DID}
\usage{
run_multiplier_bootstrap(inf.func, biters, pl = FALSE, cores = 1)
}
\description{
TAKEN DIRECTLY FROM BCALLAWAY11/DID
}
